\name{FGIST-package}
\alias{FGIST-package}
\alias{FGIST}
\docType{package}
\title{
\packageTitle{FGIST}
}
\description{
\packageDescription{FGIST}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{FGIST}
\packageIndices{FGIST}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{FGIST}

Maintainer: \packageMaintainer{FGIST}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the ~~
~~ R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
