% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GroupIST}
\alias{GroupIST}
\title{The package named GIST-fitting(Group Iterative Spectrum Thresholding)
 This package is made by Ding Fei, Zhou ya and Chen yingyi}
\usage{
GroupIST(X1, y1, lambda, Omega, w, epsilon, eta, tau0)
}
\arguments{
\item{X1}{the design matrix who has been normalized}

\item{lambda}{is the tunning parameterof the hard-ridge panalty}

\item{Omega}{is the maximum of interation}

\item{w}{is the relaxation parameter}

\item{y}{signal response who has been centered}
}
\value{
the list of betahat--the estimator of the coefficient,object function--vector of the objection function value with the interation,and interation--the interation when the objection value arrive the optimal value
}
\description{
The package named GIST-fitting(Group Iterative Spectrum Thresholding)
 This package is made by Ding Fei, Zhou ya and Chen yingyi
}
\examples{
#Generate the simulation Data
set.seed(100)
sigma<-1
N<-100
y<-c()
 for(i in 1:100){
   s<-0
   nz<-c(0.248,0.25,0.252,0.398,0.4)
   Ak<-c(2,4,3,3.5,3)
   phik<-c(pi/4,pi/6,pi/3,pi/5,pi/2)
   tn<-i
   for (fk in nz)
   {
     en<-rnorm(1,0,sigma)
     s=s+Ak*cos(2*pi*fk*tn+phik)+en
   }
   y[i]<-s
 }
 x<-c()
   for(i in 1:250){
     for(j in 1:100){
       x[j+(i-1)*100]<-cos(pi*i/250*j)
     }
   }
   for(i in 1:250){
     for(j in 1:100){
       x[25000+j+(i-1)*100]<-sin(pi*i/250*j)
     }
   }
   X<-matrix(x,100,500)
#Normalize the data X and Center the data y,so we get the using Data X1 and y1
     X1<-matrix(,100,500)
       for(i in 1:500){
         X1[,i]=(X[,i]-mean(X[,i]))/sqrt(var(X[,i]))
       }
       y1 = vector()
         for(i in 1:length(y)){
           y1[i]= y[i]-1/length(y)*sum(y)
         }
# Get the Results of the Models Using the Functon in Our Package Icluding the GroupIST and FastGroupIst
ResultGroupIST =GroupIST(X1,y1,lambda = 0.3,Omega = 10000,w = 0.5, epsilon = 0.01,eta =0.1,tau0=30)
ResultGroupIST$iteration
[1] 20
sum(ResultGroupIST$betahat!=0)
[1] 18
ResultGroupIST$`object function`[20]
[1] 13.23321
}
